==> 3-Analisis_exploratorio/datos_transaccionales/1-2-IRC.Rmd <==
---
title: "Análisis de los indicadores propuestos para la concentración de los actores en CPH"
subtitle: 'SIGMAP: Semillero de Investigación en Gestión, Mejora y Análisis de Procesos'
author: "Nicolas Arrieta y Lizeth Jerez"
date: "19/02/2021"
output:
  html_document: 
    code_folding: hide
    theme: flatly
    toc: yes
    toc_depth: 2
---

<style>
body {
text-align: justify}
</style>

# Librerias y conjunto de datos

En este caso los conjuntos de datos empleados son el resultado de las etapas de preprocesamiento.

```{r setup, message=F, warning=F}
# Librerias
sapply(
 c('dplyr', 'readr', 'ggplot2', 'gridExtra', 'lubridate',
   'knitr', 'skimr', 'reshape2', 'GGally'), 
 require, character.only = T
)

# Ajustes definidos para el documento
knitr::opts_chunk$set(message = F, warning = F, fig.align = "center")

# Carga de datos
direccion <- paste0('C:/Users/nico2/Proyectos/contratacion_publica/',
                    'riesgo-concentracion-CPH/1-1-Datasets/')
# Indicadores entidades
ind_ent <- read_csv(paste0(direccion, 'entidades/1A_entidades.csv'),
                      locale = locale(encoding = 'UTF-8'))

# Indicadores contratistas
ind_cont <- read_csv(paste0(direccion, 'contratistas/2A_contratistas.csv'),
                      locale = locale(encoding = 'UTF-8'))
```

# Análisis exploratorio

## Resumen estadístico

A continuación se presenta un resumen estadístico para cada una de las variables de los conjuntos de entidades y contratistas.

**Resumen para entidades**

```{r resumen entidad}
skim(ind_ent)
```

**Resumen para contratistas**

```{r resumen contratista}
skim(ind_cont)
```

## Análisis de correlación

### Entidades

A continuación, se presenta la correlación de las variables transaccionales para entidades. Cabe resaltar dos situaciones. En primera instancia, las variables del indice HHI y el IC4K presentan una correlación alta, debido a que entre estas 4 varaibles puede estar representada el indicador de concentración, es decir, el de target es necesario aclarar la situación. Por otro lado, hay una correlación en los indicadores que presentan modificaciones en los contratos, sin embargo, el hecho de que la mayoría de valores están alredor del cero, puede generarse este efecto, para este caso es necesario revisar variables categoricas.

```{r pt correlacion entidad, fig.height=7, fig.width=9}
correlacion <- function(data, titulo){

   data <- data %>%
    cor(use = "complete.obs")
   
   data <- round(x = data, digits = 2)
   
   data <- melt(data)
   
   ggplot(data = data, aes(x = Var1, y = Var2, 
                           fill = value)) + geom_tile() +
 theme(axis.text.x = element_text(angle = 45, vjust = 1, 
    size = 7, hjust = 1),
    axis.text.y = element_text(size = 7)) +
 scale_fill_gradient2(low = "green", 
                      high = "red", mid = "yellow", 
   midpoint = 0, limit = c(-1,1), space = "Lab", 
   name ="Correlación\nPearson") +
     labs(x = "", y = "", title = titulo,
          caption = 'Contratación pública hospitalaria en Colombia 2014-2019.')
}

correlacion(ind_ent %>% select(-nit_entidad),
            titulo = 'Correlación variables transaccionales para entidades')
```

#### IRCC {.tabset}

Entre las varaibles que representan la concentración, es decir, los indices HHI y el IC4K es necesario definir como van a estar representadas. A continuación, se presenta una matriz de gráficos de puntos para cada una de las variables, con adición a una variable que promedia todos los valores llamada 'IRCC'. De este gráfico se puede inferir varias cosas:

* La existencia de la correlación es fuerte entre las variables de igual dimensión, es decir, por número y valor. Por lo tanto, puede que también sea una opción para tomarlo como variable target.

* La variable IRCC tiene una gran representación de las 4 varaibles, debido a que la que presenta menor correlación tiene un puntaje de 0.58, lo que se puede considerar como una dependecia moderada. Por otro lado, la variable de _pval_ic4k_ tiene una correlación del 0.944, lo que se puede entender que con solo esta variable se puede componer el indice global. Sin embargo, se partira de este promedio gracias a su buena convergencia.

##### IRCC continuo

```{r pt target entidad R, fig.height=9, fig.width=9}
ind_concentracion <- c('num_HHI', 'val_HHI', 'pnum_ic4k', 'pval_ic4k')

ind_ent <- ind_ent %>% 
  mutate(IRCC = rowMeans(ind_ent[, ind_concentracion]),
         IRCCG = case_when(
                   IRCC / max(IRCC) < 0.444 ~ 'Bajo',
                   IRCC / max(IRCC) < 0.744 ~ 'Medio',
                   TRUE ~ 'Alto'))

ggpairs(ind_ent %>% 
          select(all_of(ind_concentracion), IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variables de concentración para entidades')
```

##### IRCC Agrupado

```{r pt target entidad C, fig.height=9, fig.width=9}
ggpairs(ind_ent %>% 
          select(all_of(ind_concentracion), IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variables de concentración para entidades')

# Actualización del conjunto de datos
ind_ent <- ind_ent %>% 
  select(-all_of(ind_concentracion))
```

#### {-}

Por otro lado, se creo una variable que categoriza el riesgo en tres niveles. Por ahora lo que queda en discusión es la desproporción de las clases

```{r}
kable(ind_ent %>% group_by(IRCCG) %>% count() %>% arrange(desc(n)))
```

#### Correlación con variables descriptivas {.tabset}

Es preocupante la poca correlación con la variable target y las variables descriptivas.

##### IRCC continuo

```{r pt target vs descriptivas entidad R, fig.height=9, fig.width=9}
scale2 <- function(x, na.rm = FALSE) (x - mean(x, na.rm = na.rm)) / sd(x, na.rm)

ggpairs(ind_ent %>% 
          mutate_if(is.double, scale2) %>% 
          select(contains('contratos'), contains('grupos'),
                 contains('familias'), contains('contratistas'), IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con descriptivas')
```

##### IRCC Agrupado

```{r pt target vs descriptivas entidad C, fig.height=9, fig.width=9}
ggpairs(ind_ent %>% 
          mutate_if(is.double, scale2) %>% 
          select(contains('contratos'), contains('grupos'),
                 contains('familias'), contains('contratistas'), IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con descriptivas')
```

#### {-}

#### Correlación con ind. adiciones {.tabset}

Es mejor tomar la variable de adicion como una sola, aunque también preocupa la poca correlación.

Con respecto a los indicadores de adición se observan las correlaciones altas entre las mismas variables del indicador, por lo que sería ideal agruparlas en una sola como están en la variable _adicion_.

##### IRCC continuo

```{r pt ind adiciones entidad R, fig.height=7, fig.width=7}
ind_modificacion <- c('pnum_adicion_tiempo', 'pval_adicion_tiempo',
                      'pnum_adicion_cuantia', 'pval_adicion_cuantia')

ggpairs(ind_ent %>% 
          mutate(adicion = rowMeans(ind_ent[, ind_modificacion])) %>%  
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_modificacion), adicion, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de adición')
```

##### IRCC Agrupado

```{r pt ind adiciones entidad C, fig.height=7, fig.width=7}
ggpairs(ind_ent %>% 
          mutate(adicion = rowMeans(ind_ent[, ind_modificacion])) %>%  
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_modificacion), adicion, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de adición')
```

#### {-}

Si tomamos la variable como categorica se puede apreciar un poco la distinción, pero la verdad no es tan significativo.

```{r pt target vs adiciones entidad}
# Agregación de variable agrupada
ind_ent <- ind_ent %>% 
  mutate(adicion = rowMeans(ind_ent[, ind_modificacion]))

# Puntos de cortes de la varaible adición
median_ad <- median(ind_ent$adicion)
q90_adicion <- quantile(ind_ent$adicion, probs = 0.9, names = F)

# Creación de variable categorica
ind_ent <- ind_ent %>% 
  mutate(gr_adicion =  case_when(adicion <= median_ad ~ 'Bajo',
                                 adicion <= q90_adicion ~ 'Medio',
                                 TRUE ~ 'Alto')) %>% 
  select(-all_of(ind_modificacion))

# grafica de puntos
ggplot(ind_ent, aes(x = adicion, y = IRCC, color = gr_adicion)) + 
  geom_point(alpha = 0.5, size = 1) + 
  scale_x_continuous(breaks = seq(0, 85, 5)) +
  theme_light() + 
  labs(title = 'Relación de concentración y adiciones para entidades',
       x = 'Indicador adición agrupado',
       y = 'IRCC', color = 'Grupo\nadición',
       caption = 'Contratación pública hospitalaria en Colombia 2014-2019.')
```

Al observar más en detalle la categorización en un diagrama de cajas se observa el hecho de que no hay tanta variación. Es de esperar que esta variable no salga como significativa.

```{r pt target vs grupo adicion entidad}
# grafica de cajas
ggplot(ind_ent, aes(x = gr_adicion, y = IRCC)) + geom_boxplot() +
  theme_classic() +
  labs(title = 'Relación de concentración y adiciones para entidades',
       x = 'Grupo de adición', y = 'IRCC',
       caption = 'Contratación pública hospitalaria en Colombia 2014-2019.')

# Actualización del conjuntos de datos
ind_ent <- ind_ent %>% select(-gr_adicion)
```

#### Correlación variables Contratación cerrada {.tabset}

Con respecto a las variables que demuestran la contratación cerrada, se observa uan gran proporción de entidades que tienen un porcentaje del 100%. Pero es mejor no agruparlas, debido a que la correlación esta en un término medio.

##### IRCC continuo

```{r pt target vs CC entidad R, fig.height=7, fig.width=7}
ind_cerrada <- c('pnum_cont_cerrada', 'pval_cont_cerrada')

# Agregación de variable agrupada
ind_ent <- ind_ent %>% 
  mutate(cont_cerrada = rowMeans(ind_ent[, ind_cerrada]))

# grafico de correlación
ggpairs(ind_ent %>% 
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_cerrada), cont_cerrada, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de CC')
```

##### IRCC Agrupado

```{r pt target vs CC entidad C, fig.height=7, fig.width=7}
ggpairs(ind_ent %>% 
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_cerrada), cont_cerrada, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de CC')

# Actualización del conjuntos de datos
ind_ent <- ind_ent %>% select(-cont_cerrada)
```

#### {-}

#### Correlación varaibles finales {.tabset}

Resultado de variables finales de banderas rojas

##### IRCC continuo

```{r pt target vs predictorias entidad R, fig.height=9, fig.width=9}
ggpairs(ind_ent %>% 
          select(all_of(ind_cerrada), adicion, pnum_ganadoras, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con banderas rojas')
```

##### IRCC Agrupado

```{r pt target vs predictorias entidad C, fig.height=9, fig.width=9}
ggpairs(ind_ent %>% 
          select(all_of(ind_cerrada), adicion, pnum_ganadoras, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con banderas rojas')
```

#### {-}


### Contratistas

Con respecto a los contratistas se observa un patrón similar al de las entidades, donde hay una relación alta entre los indicadores de concentración y las de adiciones en tiempo.

```{r pt correlacion contratista, fig.height=7, fig.width=9}
correlacion(ind_cont %>% select(-id_contratista_std),
            titulo = 'Correlación variables transaccionales para contratistas')
```

#### IRC {.tabset}

De igual manera que las entidades se observa una relación entre las variables de concentración, pero esta vez com más fuerza, debido a que el menor valor de correlación es de 0.621. Al promediar integar los valores de los cuatro indicadores en el IRCC se puede observar como existe una mayor correlación de los indicadores y el agrupado, lo que representa una buena elección seleccionar solamente la variable IRCC.

##### IRCC continuo

```{r pt target contratista R, fig.height=9, fig.width=9}
# Actualización de las columnas de concentración
ind_concentracion <- c('num_HHI', 'val_HHI', 'pnum_ic1k', 'pval_ic1k')

# Generación de nuevas variables
ind_cont <- ind_cont %>% 
  mutate(IRCC = rowMeans(ind_cont[, ind_concentracion]),
         IRCCG = case_when(
                   IRCC / max(IRCC) < 0.444 ~ 'Bajo',
                   IRCC / max(IRCC) < 0.744 ~ 'Medio',
                   TRUE ~ 'Alto'))
# random
set.seed(73)
sp <- sample(1:nrow(ind_cont), as.integer(nrow(ind_cont) * 0.5))

ggpairs(ind_cont[sp,] %>% 
          mutate(IRCC = rowMeans(ind_cont[sp, ind_concentracion])) %>% 
          select(all_of(ind_concentracion), IRCC),
        title = 'Correlación de variables de concentración para contratistas',
        lower = list(continuous = wrap(ggally_points, alpha = 0.3, size=0.1)))
```

##### IRCC Agrupado

```{r pt target contratista C, fig.height=9, fig.width=9}
ggpairs(ind_cont[sp,] %>% 
          mutate(IRCC = rowMeans(ind_cont[sp, ind_concentracion])) %>% 
          select(all_of(ind_concentracion), IRCCG),
        title = 'Correlación de variables de concentración para contratistas',
        lower = list(continuous = wrap(ggally_points, alpha = 0.3, size=0.1)))

# Actualización del conjunto de datos
ind_cont <- ind_cont %>%
  select(-all_of(ind_concentracion))
```

#### {-}

#### Correlación con variables descriptivas {.tabset}

Con respecto a las variables descriptivas se observa una mejor relación con la variable target, debido a que existen correlaciones por encima del 0.4, sin embargo, hay otros problemas como la existencias de outliers en variables numpericas como el número de contratos, valor total de contratos, entre otros.

##### IRCC continuo

```{r pt target vs descriptivas cont R, fig.height=10, fig.width=10}
ggpairs(ind_cont[sp,] %>% 
          mutate_if(is.double, scale2) %>% 
          select(contains('contratos'), contains('grupos'),
                 contains('familias'), contains('contratistas'),
                 num_annos_activo, num_municipios_dif, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con descriptivas')
```

##### IRCC Agrupado

```{r pt target vs descriptivas cont C, fig.height=10, fig.width=10}
ggpairs(ind_cont[sp,] %>% 
          mutate_if(is.double, scale2) %>% 
          select(contains('contratos'), contains('grupos'),
                 contains('familias'), contains('contratistas'),
                 num_annos_activo, num_municipios_dif, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con descriptivas')
```

#### {-}

#### Correlación con ind. adiciones {.tabset}

Las variables que indican las adiciones tienen el mismo comportamiento que con las entidades, por lo tanto, es mejor agruparla en una sola.

##### IRCC continuo

```{r pt ind adiciones cont R, fig.height=7, fig.width=7}
ggpairs(ind_cont[sp,] %>% 
          mutate(adicion = rowMeans(ind_cont[sp, ind_modificacion])) %>%  
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_modificacion), adicion, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de adición')
```

##### IRCC Agrupado

```{r pt ind adiciones cont C, fig.height=7, fig.width=7}
ggpairs(ind_cont[sp,] %>% 
          mutate(adicion = rowMeans(ind_cont[sp, ind_modificacion])) %>%  
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_modificacion), adicion, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de adición')
```

#### {-}



Al analizarlo por grupo se observa que se puede diferenciar un poco, pero de igual manera la nube esta muy dispersa para arrojar valores significativos.

```{r pt target vs adiciones cont}
# Agregación de variable agrupada
ind_cont <- ind_cont %>% 
  mutate(adicion = rowMeans(ind_cont[, ind_modificacion])) %>% 
  select(-all_of(ind_modificacion))

# Puntos de cortes de la varaible adición
med_adicion <- mean(ind_cont$adicion)
q90_adicion <- quantile(ind_cont$adicion, probs = 0.9, names = F)

# Creación de variable categorica
ind_cont <- ind_cont %>% 
  mutate(gr_adicion =  case_when(adicion <= med_adicion ~ 'Bajo',
                                 adicion <= q90_adicion ~ 'Medio',
                                 TRUE ~ 'Alto'))
# grafica de puntos
ggplot(ind_cont, aes(x = adicion, y = IRCC, color = gr_adicion)) + 
  geom_point(alpha = 0.5, size = 1) + 
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  theme_light() + 
  labs(title = 'Relación de concentración y adiciones para contratistas',
       x = 'Indicador adición agrupado',
       y = 'IRCC', color = 'Grupo\nadición',
       caption = 'Contratación pública hospitalaria en Colombia 2014-2019.')
```

En este caso es mejor conservar la variable continua.

```{r pt target vs grupo adicion contratista}
# grafica de cajas
ggplot(ind_cont, aes(x = gr_adicion, y = IRCC)) + geom_boxplot() +
  theme_classic() +
  labs(title = 'Relación de concentración y adiciones para contratistas',
       x = 'Grupo de adición', y = 'IRCC',
       caption = 'Contratación pública hospitalaria en Colombia 2014-2019.')

# Actualización del conjuntos de datos
ind_cont <- ind_cont %>% select(-gr_adicion)
```


#### Correlación variables Contratación cerrada {.tabset}

Con respecto a las variables de contratación cerrada es mejor considerar una unión de variables, debido a la gran correlación entre las dos variables.

##### IRCC continuo

```{r pt target vs CC agrupado cont R, fig.height=7, fig.width=7}
# Agregación de variable agrupada
ind_cont <- ind_cont %>% 
  mutate(cont_cerrada = rowMeans(ind_cont[, ind_cerrada]))

# grafico
ggpairs(ind_cont[sp,] %>% 
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_cerrada), cont_cerrada, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de CC')
```

##### IRCC Agrupado

```{r pt target vs CC agrupado cont C, fig.height=7, fig.width=7}
ggpairs(ind_cont[sp,] %>% 
          mutate_if(is.double, scale2) %>% 
          select(all_of(ind_cerrada), cont_cerrada, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con indicadores de CC')

# remover las variables separadas
ind_cont <- ind_cont %>% select(-all_of(ind_cerrada))
```

#### {-}

#### Correlación varaibles finales {.tabset}

Resultado de variables finales de banderas rojas

##### IRCC continuo

```{r pt target vs predictorias cont R, fig.height=9, fig.width=9}
ggpairs(ind_cont[sp,] %>% 
          select(cont_cerrada, adicion, pnum_entidades, IRCC),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con banderas rojas')
```

##### IRCC Agrupado

```{r pt target vs predictorias cont C, fig.height=9, fig.width=9}
ggpairs(ind_cont[sp,] %>% 
          select(cont_cerrada, adicion, pnum_entidades, IRCCG),
        lower = list(continuous = wrap(ggally_points, alpha = 0.25, size=0.1)),
        title = 'Correlación de variable target con banderas rojas')
```

#### {-}

# Hallazgos y conclusiones
* Se pudo construir la variable target a patir de los 4 indicadores de concentración más directos, que son HHI por valor y numéro, así mismo con el ICK. La variable agrupada (IRCC) mantiene significativamente la variabilidad de las cuatro variables.

* Se mejoró la configuración del conjunto de datos para entidades y contratistas, a fin de eliminar variables muy correlacionadas que puedan afectar en la inferencia.

* Hay ciertos problemas de outliers en algunas variables absolutas, en especial para el conjunto de contratistas.

* Hasta el momento las variables predictorias no tienen gran relación con la varaible target por lo que puede representar un resultado poco alentador al principio.

```{r Escritura de datos}
# Contratistas
write_csv(ind_cont, paste0(direccion, 'contratistas/2B_contratistas.csv'))
# Entidades
write_csv(ind_ent, paste0(direccion, 'entidades/1B_entidades.csv'))
```
